style(owner): Final de estilos basicos

+++ frontend/src/pages/owner/Hoteles.jsx
@@ -57,11 +57,13 @@ const Hoteles = () => {
   };
 
   const handleDeleteHotel = async (id) => {
-    try {
-      await api.delete(`/hoteles/${id}`);
-      setHoteles(hoteles.filter((hotel) => hotel.id_hotel !== id));
-    } catch (error) {
-      console.error("Error deleting hotel:", error);
+    if(confirm("¿Seguro desea eliminar este hotel?")) {
+      try {
+        await api.delete(`/hoteles/${id}`);
+        setHoteles(hoteles.filter((hotel) => hotel.id_hotel !== id));
+      } catch (error) {
+        console.error("Error deleting hotel:", error);
+      }
     }
   };
