+++ frontend/src/pages/owner/Surusales.jsx
@@ -44,11 +44,13 @@ const Sucursales = () => {
   };
 
   const handleDeleteSucursal = async (id) => {
-    try {
-      await api.delete(`/sucursales/${id}`);
-      setSucursales(sucursales.filter((sucursal) => sucursal.id_sucursal !== id));
-    } catch (error) {
-      console.error("Error deleting sucursal:", error);
+    if(confirm("¿Seguro desea eliminar esta sucursal?")) {
+      try {
+        await api.delete(`/sucursales/${id}`);
+        setSucursales(sucursales.filter((sucursal) => sucursal.id_sucursal !== id));
+      } catch (error) {
+        console.error("Error deleting sucursal:", error);
+      }
     }
   };
